1st version of a responsive design - mobile first
I adjusted my 1st project into a responsive design with the focus on smartphone sizes. 

The workbook template helped me a lot with the definition of the general layouts according to the breakpoints. Unfortunately, I did not have that much time to implement it accordingly. I got stuck on the layout of the menu on the welcome page (index.html). I want the basic layout to consist of pictures with a hover effect changing to text - more precisely changing to a link to the corresponding page. That didn't work out so far - I will need to work on this more in detail.