Add files via upload
# Project Name - Flower Shop / Subscription Service

For project #2, I reused the code I've written for the first project. 
Project #2 focusses on the product category page '/our-flowers.html'.
Therefore, please ignore the other pages. 

## The problem

This page is intended to show a variety of products, in 2 different styles. 

The mobile version contains product cards that are aligned in columns, with an image on the left and further details on the right. 

The tablet/desktop version displays multiple products in a row, with the product details below the image. 

In addition to that, the background color and borders of the product cards change when adjusting the browser size. 

Further, the full navigation bar and extended headings only show in the desktop version. 
The mobile version shows a hamburger navigation and only a short heading. 

## View it live

Have you deployed your project somewhere? Be sure to include the link to the deployed project so that the viewer can click around and see what it's all about.